+++ packages/web/src/features/chat/components/chatThread/chatThread.tsx
@@ -269,6 +269,7 @@ export const ChatThread = ({
                                 return (
                                     <Fragment key={index}>
                                         <ChatThreadListItem
+                                            index={index}
                                             chatId={chatId}
                                             userMessage={userMessage}
                                             assistantMessage={assistantMessage}
