+++ packages/web/src/features/chat/constants.ts
@@ -1,3 +1,4 @@
+import { SBChatMessagePart } from "./types";
 
 export const FILE_REFERENCE_PREFIX = '@file:';
 export const FILE_REFERENCE_REGEX = new RegExp(
@@ -14,3 +15,13 @@ export const toolNames = {
     findSymbolReferences: 'findSymbolReferences',
     findSymbolDefinitions: 'findSymbolDefinitions',
 } as const;
+
+// These part types are visible in the UI.
+export const uiVisiblePartTypes: SBChatMessagePart['type'][] = [
+    'reasoning',
+    'text',
+    'tool-searchCode',
+    'tool-readFiles',
+    'tool-findSymbolDefinitions',
+    'tool-findSymbolReferences',
+] as const;
