+++ packages/web/src/features/chat/useExtractReferences.ts
@@ -1,39 +1,36 @@
 'use client';
 
 import { useMemo } from "react";
-import { SBChatMessage, FileReference } from "./types";
+import { FileReference } from "./types";
 import { FILE_REFERENCE_REGEX } from "./constants";
 import { createFileReference } from "./utils";
+import { TextUIPart } from "ai";
 
-export const useExtractReferences = (message?: SBChatMessage) => {
+export const useExtractReferences = (part?: TextUIPart) => {
     return useMemo(() => {
+        if (!part) {
+            return [];
+        }
+
         const references: FileReference[] = [];
 
-        message?.parts.forEach((part) => {
-            switch (part.type) {
-                case 'text':
-                case 'reasoning': {
-                    const content = part.text;
-                    FILE_REFERENCE_REGEX.lastIndex = 0;
+        const content = part.text;
+        FILE_REFERENCE_REGEX.lastIndex = 0;
 
-                    let match;
-                    while ((match = FILE_REFERENCE_REGEX.exec(content ?? '')) !== null && match !== null) {
-                        const [_, repo, fileName, startLine, endLine] = match;
+        let match;
+        while ((match = FILE_REFERENCE_REGEX.exec(content ?? '')) !== null && match !== null) {
+            const [_, repo, fileName, startLine, endLine] = match;
 
-                        const fileReference = createFileReference({
-                            repo: repo,
-                            path: fileName,
-                            startLine,
-                            endLine,
-                        });
+            const fileReference = createFileReference({
+                repo: repo,
+                path: fileName,
+                startLine,
+                endLine,
+            });
 
-                        references.push(fileReference);
-                    }
-                    break;
-                }
-            }
-        });
+            references.push(fileReference);
+        }
 
         return references;
-    }, [message]);
+    }, [part]);
 };
